+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## node-red-contrib-letsfiware-NGSI v0.10.0-next
 
+-  ADD OpenWeatherMap to NGSI custom node (#107)
 -  UPDATE documentation (#105)
 
 ## node-red-contrib-letsfiware-NGSI v0.10.0 - 21 February, 2023